pyproject.toml
[tool.poetry]
name = "binance-cli"
version = "0.1.0"
description = ""
authors = ["Brian Li <brian@brianli.com>"]

[tool.poetry.scripts]
binance = "binance_cli.main:app"

[tool.poetry.dependencies]
python = "^3.10"
python-binance = "^1.0.16"
typer = "^0.6.1"
rich = "^12.5.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
